requirements.txt
ipython
python-dotenv
scrapy
